% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datach.R
\name{datach}
\alias{datach}
\title{datach}
\usage{
datach(data, x)
}
\arguments{
\item{data}{A data entry is required.}

\item{x}{The 3 categorical variables that you make matches for.}
}
\value{
A list with data.
}
\description{
Generate new data and define the data.
}
