% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm3datalist.R
\name{pm3datalist}
\alias{pm3datalist}
\title{pm3datalist}
\usage{
pm3datalist(data, x, y, covs, factor = NULL)
}
\arguments{
\item{data}{A data entry is required.}

\item{x}{The 3 categorical variables that you make matches for.}

\item{y}{Your result variable.}

\item{covs}{Enter the relevant covariates.}

\item{factor}{Define categorical variables.}
}
\value{
A data.
}
\description{
Identification and formatting of data.
}
