% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm3fit.R
\name{pm3fit}
\alias{pm3fit}
\title{pm3fit}
\arguments{
\item{data}{A data entry is required.}

\item{x}{The 3 categorical variables that you make matches for.}

\item{y}{Your result variable.}

\item{covs}{Enter the relevant covariates.}

\item{factor}{Define categorical variables.}
}
\value{
A list with data.
}
\description{
Generate propensity scores and generate the data to be matched.
}
