% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{score_pm3}
\alias{score_pm3}
\title{Calculate LOTTR or TTTR Metric Score}
\usage{
score_pm3(infile, metric = "LOTTR", verbose = FALSE)
}
\arguments{
\item{infile}{Path to the input file with travel time readings}

\item{metric}{"LOTTR" or "TTTR"}

\item{verbose}{Provide diagnostic output}
}
\value{
A data.table of TMC scores
}
\description{
Calculate LOTTR / TTTR given a RITIS export of travel time data
input file must have header and format: 
tmc_code,measurement_tstamp,travel_time_seconds
 e.g. 
 > tmc_code,measurement_tstamp,travel_time_seconds
 > 116-04379,2019-01-01 00:00:00,44.78
 > 116-04379,2019-01-01 00:15:00,46.69
}
