% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm3.R
\docType{package}
\name{pm3}
\alias{pm3}
\title{PM3 Tools: A package for computating the notorious bar statistic.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

